move sensor helper inside standard provider
